Update build instructions
Changes include:
- Explicitly require root permissions on Linux for disk creation.
- Exit both create and mount scripts on Linux if root access is not available.
- Update build instructions in README.md.

+++ scripts/createdisk.sh
@@ -2,6 +2,11 @@
 # createdisk.sh: creates a disk image with GRUB installed on it.
 # --------------------------------------------------------------
 
+if [[ ! $(id -u) = 0 ]]; then
+    printf "[mkdisk] cannot run without root.\n"
+    exit 1
+fi
+
 dd if=/dev/zero of=build/disk.img bs=512 count=131072
 sfdisk build/disk.img < scripts/layout.sfdisk
 
@@ -13,5 +18,4 @@ mkdir -p build/mnt
 mount $loopgrub build/mnt
 grub-install --root-directory=$(pwd)/build/mnt --no-floppy --modules="normal part_msdos ext2 multiboot" $loopdisk
 umount build/mnt
-rmdir build/mnt
 losetup --detach $loopdisk $loopgrub
